+++ docs/.vitepress/demos/switch/basic/Demo.vue
@@ -1,0 +1,23 @@
+<script setup lang="ts">
+import { VcSwitch } from '@wisemen/vue-core'
+import { ref } from 'vue'
+
+const model = ref<boolean>(false)
+const model2 = ref<boolean>(false)
+</script>
+
+<template>
+  <div>
+    <VcSwitch
+      v-model="model"
+      label="I want to receive marketing emails"
+      hint="You can unsubscribe at any time."
+    />
+    <div class="w-72 border-t border-primary my-3" />
+    <VcSwitch
+      v-model="model2"
+      label="You can also use a smaller switch"
+      size="sm"
+    />
+  </div>
+</template>