+++ packages/components/src/components/index.ts
@@ -31,6 +31,10 @@ export type VcTableProps<TSchema, TFilters> = TableProps<TSchema, TFilters>
 export type VcTextFieldProps = TextFieldProps
 export type VcTooltipProps = TooltipProps
 
+// Animated Number
+export { default as VcAnimatedNumber } from '@/components/animated-number/AnimatedNumber.vue'
+export { default as VcAnimatedNumberGroup } from '@/components/animated-number/AnimatedNumberGroup.vue'
+
 // Autocomplete
 export { default as VcAddressAutocomplete } from '@/components/autocomplete/AddressAutocomplete.vue'
 export { default as VcAutocomplete } from '@/components/autocomplete/Autocomplete.vue'