+++ packages/components/src/components/switch/Switch.vue
@@ -39,6 +39,21 @@ const emit = defineEmits<{
   focus: [event: Event]
 }>()
 
+defineSlots<{
+  /**
+   * Can be used to render you own hint and errors.
+   */
+  bottom: () => null
+  /**
+   * Can be used to render your own errors.
+   */
+  error: () => null
+  /**
+   * Can be used to render your own hint.
+   */
+  hint: () => null
+}>()
+
 const model = defineModel<boolean>({
   required: true,
 })
@@ -80,25 +95,15 @@ const thumbClasses = computed<string>(() => switchStyle.thumb({
   isHovered: isHovered.value,
 }))
 
-const labelClasses = computed<string>(() => switchStyle.label({
-  hasError: hasError.value,
-  isChecked: isChecked.value,
-  isDisabled: isDisabled.value,
-  isFocused: isFocused.value,
-  isHovered: isHovered.value,
-}))
-
-const hintClasses = computed<string>(() => switchStyle.hint({
-  hasError: hasError.value,
-  isChecked: isChecked.value,
-  isDisabled: isDisabled.value,
-  isFocused: isFocused.value,
-  isHovered: isHovered.value,
-}))
-
-const errorClasses = computed<string>(() => switchStyle.error())
-
 const iconClasses = computed<string>(() => switchStyle.icon({
+  hasError: hasError.value,
+  isChecked: isChecked.value,
+  isDisabled: isDisabled.value,
+  isFocused: isFocused.value,
+  isHovered: isHovered.value,
+}))
+
+const labelClasses = computed<string>(() => switchStyle.label({
   hasError: hasError.value,
   isChecked: isChecked.value,
   isDisabled: isDisabled.value,
@@ -195,10 +200,10 @@ function onBlur(e: Event): void {
 
       <InputFieldLabel
         v-if="props.label !== null"
-        :class="labelClasses"
         :is-required="props.isRequired"
         :for="inputId"
         :label="props.label"
+        :class="labelClasses"
       />
 
       <!-- Spacer element for grid -->
@@ -214,7 +219,6 @@ function onBlur(e: Event): void {
               <slot name="error">
                 <InputFieldError
                   :errors="props.errors"
-                  :class="errorClasses"
                   :input-id="inputId"
                 />
               </slot>
@@ -225,7 +229,6 @@ function onBlur(e: Event): void {
                 <InputFieldHint
                   :input-id="inputId"
                   :hint="props.hint"
-                  :class="hintClasses"
                 />
               </slot>
             </div>