+++ packages/components/src/components/table/TableCell.vue
@@ -8,6 +8,13 @@ const props = withDefaults(defineProps<{
 }>(), {
   isPrimaryCell: false,
 })
+
+defineSlots<{
+  /**
+   * Use this to render the table cell data.
+   */
+  default: () => void
+}>()
 </script>
 
 <template>