+++ packages/components/src/components/tabs/TabsContent.vue
@@ -2,7 +2,17 @@
 import { TabsContent } from 'reka-ui'
 
 const props = defineProps<{
+  /**
+ * A unique value that associates the content with a trigger.
+ */
   value: string
+}>()
+
+defineSlots<{
+  /**
+   * The content to display in the tab.
+   */
+  'default': () => void
 }>()
 </script>